+++ ECS/plainECS.sln
@@ -36,11 +36,9 @@ Global
 		{600092B0-1DA9-4C65-91E5-271BE2278860}.Debug|x86.ActiveCfg = Debug|Win32
 		{600092B0-1DA9-4C65-91E5-271BE2278860}.Debug|x86.Build.0 = Debug|Win32
 		{600092B0-1DA9-4C65-91E5-271BE2278860}.Lib_d|x64.ActiveCfg = Release|x64
-		{600092B0-1DA9-4C65-91E5-271BE2278860}.Lib_d|x64.Build.0 = Release|x64
 		{600092B0-1DA9-4C65-91E5-271BE2278860}.Lib_d|x86.ActiveCfg = Release|Win32
 		{600092B0-1DA9-4C65-91E5-271BE2278860}.Lib_d|x86.Build.0 = Release|Win32
 		{600092B0-1DA9-4C65-91E5-271BE2278860}.Lib|x64.ActiveCfg = Release|x64
-		{600092B0-1DA9-4C65-91E5-271BE2278860}.Lib|x64.Build.0 = Release|x64
 		{600092B0-1DA9-4C65-91E5-271BE2278860}.Lib|x86.ActiveCfg = Release|Win32
 		{600092B0-1DA9-4C65-91E5-271BE2278860}.Lib|x86.Build.0 = Release|Win32
 		{600092B0-1DA9-4C65-91E5-271BE2278860}.Release|x64.ActiveCfg = Release|x64
