+++ ECS/plainECS/src/ecsBaseEvent.cpp
@@ -1,0 +1,8 @@
+#include "ecsBaseEvent.h"
+
+using namespace ecs;
+
+
+// Make id counters start at 1, since 0=error
+TypeID BaseEvent::typeIDCounter(1);
+//IDGenerator<TypeID> BaseEvent::typeIDGenerator(1);
